requirements.txt
feedparser
openai
requests
PyYAML
